Rename wrapper folders.
