+++ wrappers/Python/MANIFEST.in
